pyproject.toml
[project]
name = "rag-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "google-genai>=1.16.1",
    "llama-cloud-services>=0.6.23",
    "python-dotenv>=1.1.0",
    "redis>=6.1.0",
]
